rag_demo/backend/requirements.txt
# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Streamlit UI
streamlit>=1.28.0

# PDF processing
pypdf>=3.17.0
pdfplumber>=0.10.0
python-docx>=1.1.0  # Optional, for DOCX support
reportlab>=4.0.0  # PDF generation with Unicode support

# Vector database
chromadb>=0.5.0

# HTTP requests
requests>=2.31.0

# OpenAI API
openai>=1.0.0

# Data models
pydantic>=2.5.0

# Environment variables
python-dotenv>=1.0.0

# Text processing
tiktoken>=0.5.0

# Utilities
numpy>=1.26.0  # Python 3.12 support
